Implements dependency injection
Introduces a dependency container to manage and inject services like ApiService and ImageService.

This change promotes loose coupling and testability by providing dependencies to the ListContactsViewController and ContactViewModel.

Removes the hardcoded ApiService initialization from the ListContactsFactory and moves it to the DependenciesContainer.

This facilitates easier configuration and mocking of dependencies for different environments and testing scenarios.